requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
Flask-SQLAlchemy
WTForms
